refactor: location queries
