+++ lib/db/schema/location.ts
@@ -23,3 +23,5 @@ export const InsertLocation = z.object({
   lat: z.coerce.number().min(-90).max(90),
   long: z.coerce.number().min(-180).max(180),
 });
+
+export type InsertLocation = z.infer<typeof InsertLocation>;
